C++ Solution for problem "Container with most water" Signed-off-by: Shubham Chaudhary <shubhamchaudhary632001@gmail.com>

+++ Programming/C++/ContainerWithMostWater/container_with_most_water_solution_cpp.cpp
@@ -24,7 +24,7 @@ int main()
     int n;
     cin>>n;
     
-    vector<ll> a(n);
+    vector<int> a(n);
     for(int i=0; i<n; i++)
     {
         cin>>a[i];
